feat: support changing config at runtime
Instead of only doing so via CLI args. Currently starts a new board with
the new options any time config is accessed and you hit done. Should be
able to update just the tick time by changing tick time in the config
and then hitting ESC to directly bring up the menu and then going to
Play to view the board.